Fix: Return error if same name file exists when creating

+++ src/main.rs
@@ -99,8 +99,13 @@ async fn read_item(Path(file_name): Path<String>) -> Result<Json<Item>, Error> {
 #[debug_handler]
 async fn create_item(new_file_name: String) -> Result<impl IntoResponse, Error> {
     println!("[CREATE] {}", new_file_name);
-    std::fs::File::create(create_path(&new_file_name))?;
-    Ok(new_file_name.into_response())
+    let path = create_path(&new_file_name);
+    if path.exists() {
+        Err(Error::SameName)
+    } else {
+        std::fs::File::create(create_path(&new_file_name))?;
+        Ok(new_file_name.into_response())
+    }
 }
 
 #[debug_handler]
